requirements_simple.txt
# Bot Trading v10 Enterprise - Dependencias Simplificadas
# =======================================================

# Core Dependencies
python-telegram-bot==20.7
aiohttp==3.9.1
aiofiles==23.2.1
aiosqlite==0.19.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Machine Learning
torch>=2.0.0
pytorch-lightning>=2.0.0

# Trading & Finance
ccxt>=4.0.0
ta>=0.10.0
yfinance>=0.2.0
websockets>=12.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.13.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.5.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.0.0
colorlog>=6.8.0

# Security & Encryption
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Visualization
matplotlib>=3.8.0
plotly>=5.17.0
dash>=2.14.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# System Dependencies
psutil>=5.9.0
